Add verify (#5)
Add verify command which:

does sync_shards for every DB
does an _all_docs request (with limit 0
checks view indexes (returning results on a stale request is 👍🏻 )
Caveat for views: if the view has no docs to index (none of the docs emit in the view), we will get false failures, so on no doc views we just display a warning. Wording of the warning can definitely be improved.